Fix: 직원 리스트 페이지에서 직원 클릭시 페이지 이동이 두 번되는 버그 해결 - 직원 리스트 페이지 EmployeeListTable.js 에  onClickTableRow 함수를 수정. -  this.container에 클릭 이벤트 함수를 제거하고, 테이블로우에 클릭 이벤트 함수를 추가함
